requirements.txt
pandas>=2.0
pydantic>=2.6
jinja2>=3.1
PyYAML>=6.0
